+++ chroma/colors/errors.py
@@ -1,0 +1,7 @@
+class IncompatibleColorError(Exception):
+    """
+    This is thrown when two colors are of different and/or incompatible types.
+    Please make sure that the colors are of similar types before performing the
+    operation.
+    """
+    pass