+++ chroma/utils/theme.py
@@ -2,6 +2,7 @@ from pathlib import Path
 
 from lupa import LuaRuntime
 
+from chroma.colors.methods import darken, desaturate, lighten, saturate
 from chroma.logger import Logger
 
 from .paths import cache_dir, chroma_dir
@@ -9,32 +10,72 @@ from .tools import backup, to_dict
 
 logger = Logger.get_logger()
 
-DEFAULT_STATE: dict = {"use_generated": True}
+DEFAULT_STATE: dict = {
+    "use_generated": True,
+
+    "darken": darken,
+    "lighten": lighten,
+    "saturate": saturate,
+    "desaturate": desaturate,
+
+    # "chroma": {
+    #     "colors": {
+    #         "methods": {
+    #             "darken": darken,
+    #             "lighten": lighten,
+    #             "saturate": saturate,
+    #             "desaturate": desaturate,
+    #         }
+    #     }
+    # },
+}
 
 
+# TODO: add support for lists/arrays
 def sanitize_python(state: dict = dict(), **kwargs) -> dict:
     state.update(kwargs)
-    out = dict()
+
+    def sanitize_value(value):
+        if isinstance(value, bool):
+            return "true" if value else "false"
+        elif isinstance(value, (int, float)):
+            return str(value)
+        elif isinstance(value, str):
+            return f'"{value}"'
+        elif isinstance(value, dict):
+            return sanitize_python(value)
+        elif value is None:
+            return "nil"
+        elif callable(value):
+            return value
+        else:
+            raise ValueError(f"Value of type {type(value)} ({value}) is unsupported")
+
+    sanitized = dict()
     for name, value in state.items():
-        # Sanitize the boolean type to match lua expectations
-        if type(value) is bool:
-            value = "true" if value else "false"
-
-        # Output the sanitized state
-        out[name] = value
-    return out
+        sanitized[name] = sanitize_value(value)
+    return sanitized
 
 
+# TODO: add tabled global methods
 def runtime(state: dict | None = DEFAULT_STATE):
     runtime = LuaRuntime(unpack_returned_tuples=True)  # pyright: ignore
     runtime.execute(f"package.path = package.path .. ';{chroma_dir().parent}/?.lua'")
     runtime.execute(f"package.path = package.path .. ';{cache_dir().parent}/?.lua'")
 
+    def register_state(d):
+        for key, value in d.items():
+            if isinstance(value, dict):
+                register_state(value)
+            elif callable(value):
+                logger.debug(f"Setting function '{key}'")
+                runtime.globals()[key] = value
+            else:
+                logger.debug(f"Setting state '{key}' = '{value}'")
+                runtime.execute(f"{key} = {value}")
+
     if state is not None:
-        state = sanitize_python(state)
-        for name, value in state.items():
-            runtime.execute(f"{name} = {value}")
-            logger.debug(f"Setting state '{name}' = '{value}'")
+        register_state(sanitize_python(state))
 
     return runtime